Allow for germline sequence lookup in masking
I modified the argparse to make the mask option a string, and one of
several options.
Notably, one of the options is to look up the sequence of the germline
from which this sequence is descended and put that sequence into the
GERMLINE_GAP_D_MASK field.
Currently, no method is implemented to proprerly mask the D region of
the germline sequence, so the unmasked germline sequence is used.